MAT-7465 refactoring and test cases
