+++ src/CqlBuilderPanel/Includes/Search/LibrarySearch.tsx
@@ -1,12 +1,12 @@
 import React, { useReducer, useState } from "react";
-import ExpandingSection from "../../common/ExpandingSection";
-import Search from "./Search";
+import ExpandingSection from "../../../common/ExpandingSection";
+import Search from "../Search";
 import useCqlLibraryServiceApi, {
   CqlLibrary,
-} from "../../api/useCqlLibraryServiceApi";
+} from "../../../api/useCqlLibraryServiceApi";
 import { Toast } from "@madie/madie-design-system/dist/react";
-import Results from "./Results/Results";
-import toastReducer from "../../common/ToastReducer";
+import Results from "../Results/Results";
+import toastReducer from "../../../common/ToastReducer";
 
 export interface LibraryTabContentProps {
   canEdit: boolean;
