+++ src/api/useCqlLibraryServiceApi.test.ts
@@ -24,13 +24,13 @@ const cqlLibraries = [
     id: "1",
     cqlLibraryName: "Test_1",
     version: "1.1.111",
-    owner: "John",
+    librarySet: { owner: "John" },
   },
   {
     id: "2",
     cqlLibraryName: "Test_0",
     version: "1.1.110",
-    owner: "Jane",
+    librarySet: { owner: "John" },
   },
 ] as Array<CqlLibrary>;
 
@@ -48,10 +48,11 @@ describe("CqlLibraryServiceApi Tests", () => {
     mockedAxios.get.mockImplementation((url) =>
       Promise.resolve({ data: cqlLibraries })
     );
-    const response = await cqlLibraryServiceApi.fetchVersionedCqlLibraries(
-      "test",
-      "QDM"
-    );
+    const response =
+      await cqlLibraryServiceApi.fetchVersionedCqlLibrariesBySearchTermAndModel(
+        "test",
+        "QDM"
+      );
     expect(response.length).toBe(2);
     expect(response[0].cqlLibraryName).toEqual(cqlLibraries[0].cqlLibraryName);
     expect(response[0].version).toEqual(cqlLibraries[0].version);
@@ -64,9 +65,32 @@ describe("CqlLibraryServiceApi Tests", () => {
       Promise.reject({ status: 400 })
     );
     try {
-      await cqlLibraryServiceApi.fetchVersionedCqlLibraries("test", "");
+      await cqlLibraryServiceApi.fetchVersionedCqlLibrariesBySearchTermAndModel(
+        "test",
+        ""
+      );
     } catch (e) {
       expect(e.message).toEqual(fetchVersionedLibrariesErrorMessage);
     }
+  });
+
+  it("Should return versioned cql library for given name, version and model", async () => {
+    mockedAxios.get.mockImplementation((url) =>
+      Promise.resolve({ data: cqlLibraries[0], status: 200 })
+    );
+    const response =
+      await cqlLibraryServiceApi.getVersionedCqlLibraryByNameVersionAndModel(
+        "Test_1",
+        "1.1.111",
+        "QDM"
+      );
+    expect(response.status).toEqual(200);
+    expect(response.data.cqlLibraryName).toEqual(
+      cqlLibraries[0].cqlLibraryName
+    );
+    expect(response.data.version).toEqual(cqlLibraries[0].version);
+    expect(response.data.librarySet.owner).toEqual(
+      cqlLibraries[0].librarySet.owner
+    );
   });
 });
